generate-changes: get index information from releases
Instead of maintaining a manual list of indexes that needs to be kept
up-to-date, use the information from alpinelinux.org/releases.json to
determine what indexes to check.

This file is updated after every release, so we do not need to remember
to update the list of indexes for the mirror-status to show.

What still can be improved is that the information in releases.json is
also used to determine what arches and branches to show.

Fixes alpine/infra/mirror-status#2